requirements.txt
alembic==1.0.0
asn1crypto==0.24.0
attrs==18.2.0
Automat==0.7.0
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
click==6.7
constantly==15.1.0
cryptography==1.9
cssselect==1.0.3
decorator==4.3.0
Django==2.1.7
django-cors-headers==2.5.2
django-mongodb-engine==0.6.0
djangorestframework==3.9.2
djangotoolbox==1.8.0
Flask==0.12.2
Flask-Cors==3.0.3
Flask-Migrate==2.2.1
Flask-SQLAlchemy==2.3.2
GDAL==2.2.4
GeoAlchemy2==0.5.0
hyperlink==18.0.0
idna==2.6
incremental==17.5.0
ipython-genutils==0.2.0
itsdangerous==0.24
Jinja2==2.10
jsonschema==2.6.0
jupyter-core==4.4.0
lxml==4.3.1
Mako==1.0.7
MarkupSafe==1.0
nbformat==4.4.0
numpy==1.14.2
OWSLib==0.16.0
panda==0.3.1
parsel==1.5.1
Pillow==5.2.0
plotly==2.5.1
psycopg2==2.7.4
psycopg2-binary==2.7.5
pyasn1==0.4.5
pyasn1-modules==0.2.4
pycparser==2.18
PyDispatcher==2.0.5
Pygments==2.2.0
PyHamcrest==1.9.0
PyJWT==1.6.1
pymongo==3.7.2
PyPDF2==1.26.0
pyproj==1.9.5.1
python-dateutil==2.7.2
python-editor==1.0.3
python-http-client==3.1.0
pytz==2018.4
PyYAML==3.12
queuelib==1.5.0
reportlab==3.5.6
requests==2.18.4
Scrapy==1.6.0
sendgrid==5.6.0
service-identity==18.1.0
Shapely==1.6.1
six==1.11.0
SQLAlchemy==1.2.6
sqlparse==0.2.4
traitlets==4.3.2
Twisted==18.9.0
urllib3==1.22
virtualenv==15.2.0
w3lib==1.20.0
Werkzeug==0.14.1
zope.interface==4.6.0
